grrr

+++ sieve/fake_rels.cpp
@@ -480,7 +480,7 @@ vector<index_t> all_comp_sq_2(uint64_t q0, uint64_t q1, uint64_t qfac_min,
             }
         }
     }
-    fprintf(stderr, "Got %lu 2-composite sq\n", list.size()>>1);
+    fprintf(stderr, "Got %zu 2-composite sq\n", (size_t)(list.size()>>1));
     return list;
 }
 
@@ -524,7 +524,7 @@ vector<index_t> all_comp_sq_3(uint64_t q0, uint64_t q1, uint64_t pos_min,
             }
         }
     }
-    fprintf(stderr, "Got %lu 3-composite sq\n", list.size()/3);
+    fprintf(stderr, "Got %zu 3-composite sq\n", (size_t)(list.size()/3));
     return list;
 }
